Funcionalidade Mostrar perguntas de um evento ja salvo no formulario de editar corrigida.

+++ BackEnd/service/src/main/java/com/basis/sge/service/servico/EventoServico.java
@@ -1,16 +1,15 @@
 package com.basis.sge.service.servico;
 
-import com.basis.sge.service.dominio.Evento;
-import com.basis.sge.service.dominio.EventoPergunta;
-import com.basis.sge.service.dominio.PreInscricao;
-import com.basis.sge.service.dominio.Usuario;
+import com.basis.sge.service.dominio.*;
 import com.basis.sge.service.mensagem.EmailMensagem;
 import com.basis.sge.service.repositorio.*;
 import com.basis.sge.service.servico.dto.EventoDTO;
 import com.basis.sge.service.servico.dto.EventoListagemDTO;
+import com.basis.sge.service.servico.dto.PerguntaDTO;
 import com.basis.sge.service.servico.exception.RegraNegocioException;
 import com.basis.sge.service.servico.mapper.EventoListagemMapper;
 import com.basis.sge.service.servico.mapper.EventoMapper;
+import com.basis.sge.service.servico.mapper.PerguntaMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
@@ -29,6 +28,8 @@ public class EventoServico {
 
     private final TipoEventoRepositorio tipoEventoRepositorio;
 
+    private final PerguntaMapper perguntaMapper;
+
     private final EventoMapper eventoMapper;
 
     private final InscricaoRepositorio inscricaoRepositorio;
@@ -37,9 +38,17 @@ public class EventoServico {
 
     private final EventoListagemMapper eventoListagemMapper;
 
+    public List<PerguntaDTO> listarPerguntaEvento(Integer id){
+        List<PerguntaDTO> perguntas = new ArrayList<PerguntaDTO>();
+        List<EventoPergunta> eventoPerguntaList= eventoPerguntaRepositorio.findAllByEventoId(id);
+        eventoPerguntaList.forEach((eventoPergunta) -> {
+            perguntas.add(perguntaMapper.toDto(eventoPergunta.getPergunta()));
+        });
+        return perguntas;
+    }
+
     public List<EventoListagemDTO> listar() {
         List<Evento> listaEvento = eventoRepositorio.findAll();
-
         return eventoListagemMapper.toDto(listaEvento);
     }
 
@@ -83,6 +92,10 @@ public class EventoServico {
     public void deletar(Integer id) {
         validaIdEvento(id);
         Evento evento = eventoRepositorio.findById(id).orElseThrow(() -> new RegraNegocioException("Evento não existe"));
+        List<PreInscricao> inscricao = inscricaoRepositorio.findAllByEvento(evento);
+        if(!inscricao.isEmpty()){
+            throw new RegraNegocioException("Este evento possui inscrições");
+        }
         notificarInscritos(evento.getTitulo(),id,1);
         eventoRepositorio.deleteById(id);
     }
